+++ src/config/config_generator.cc
@@ -645,6 +645,7 @@ void ConfigGenerator::generateImportOptions(
         { ConfigVal::IMPORT_LIBOPTS_EXIF_METADATA_TAGS_LIST, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_EXIF_CHARSET, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_EXIF_ENABLED, ConfigLevel::Example },
+        { ConfigVal::IMPORT_LIBOPTS_EXIF_CONTENT_LIST, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_EXIF_COMMENT_LIST, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_EXIF_COMMENT_ENABLED, ConfigLevel::Example },
 #endif
@@ -653,6 +654,7 @@ void ConfigGenerator::generateImportOptions(
         { ConfigVal::IMPORT_LIBOPTS_EXIV2_METADATA_TAGS_LIST, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_EXIV2_CHARSET, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_EXIV2_ENABLED, ConfigLevel::Example },
+        { ConfigVal::IMPORT_LIBOPTS_EXIV2_CONTENT_LIST, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_EXIV2_COMMENT_LIST, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_EXIV2_COMMENT_ENABLED, ConfigLevel::Example },
 #endif
@@ -661,6 +663,7 @@ void ConfigGenerator::generateImportOptions(
         { ConfigVal::IMPORT_LIBOPTS_ID3_METADATA_TAGS_LIST, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_ID3_CHARSET, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_ID3_ENABLED, ConfigLevel::Example },
+        { ConfigVal::IMPORT_LIBOPTS_ID3_CONTENT_LIST, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_ID3_COMMENT_LIST, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_ID3_COMMENT_ENABLED, ConfigLevel::Example },
 #endif
@@ -669,6 +672,7 @@ void ConfigGenerator::generateImportOptions(
         { ConfigVal::IMPORT_LIBOPTS_FFMPEG_METADATA_TAGS_LIST, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_FFMPEG_CHARSET, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_FFMPEG_ENABLED, ConfigLevel::Example },
+        { ConfigVal::IMPORT_LIBOPTS_FFMPEG_CONTENT_LIST, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_FFMPEG_COMMENT_LIST, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_FFMPEG_COMMENT_ENABLED, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_FFMPEG_ARTWORK_ENABLED, ConfigLevel::Example },
@@ -679,6 +683,7 @@ void ConfigGenerator::generateImportOptions(
         { ConfigVal::IMPORT_LIBOPTS_MKV_METADATA_TAGS_LIST, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_MKV_CHARSET, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_MKV_ENABLED, ConfigLevel::Example },
+        { ConfigVal::IMPORT_LIBOPTS_MKV_CONTENT_LIST, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_MKV_COMMENT_ENABLED, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_MKV_COMMENT_LIST, ConfigLevel::Example },
 #endif
@@ -687,6 +692,7 @@ void ConfigGenerator::generateImportOptions(
         { ConfigVal::IMPORT_LIBOPTS_WAVPACK_METADATA_TAGS_LIST, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_WAVPACK_CHARSET, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_WAVPACK_ENABLED, ConfigLevel::Example },
+        { ConfigVal::IMPORT_LIBOPTS_WAVPACK_CONTENT_LIST, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_WAVPACK_COMMENT_ENABLED, ConfigLevel::Example },
         { ConfigVal::IMPORT_LIBOPTS_WAVPACK_COMMENT_LIST, ConfigLevel::Example },
 #endif