+++ src/metadata/taglib_handler.cc
@@ -71,6 +71,12 @@
 #include <dsffile.h>
 #include <mp4itemfactory.h>
 #include <wavfile.h>
+
+#if TAGLIB_MINOR_VERSION >= 2
+#include <matroskaattachedfile.h>
+#include <matroskaattachments.h>
+#include <matroskafile.h>
+#endif
 #endif
 
 class GerberaTagLibDebugListener : public TagLib::DebugListener {
@@ -94,6 +100,7 @@ GerberaTagLibDebugListener GerberaTagLibDebugListener::grbListener;
 TagLibHandler::TagLibHandler(const std::shared_ptr<Context>& context)
     : MediaMetadataHandler(context,
           ConfigVal::IMPORT_LIBOPTS_ID3_ENABLED,
+          ConfigVal::IMPORT_LIBOPTS_ID3_CONTENT_LIST,
           ConfigVal::IMPORT_LIBOPTS_ID3_METADATA_TAGS_LIST,
           ConfigVal::IMPORT_LIBOPTS_ID3_AUXDATA_TAGS_LIST,
           ConfigVal::IMPORT_LIBOPTS_ID3_COMMENT_ENABLED,
@@ -112,7 +119,7 @@ bool TagLibHandler::isSupported(
     const std::string& mimeType,
     ObjectType mediaType)
 {
-    return contentType == CONTENT_TYPE_MP3 || (contentType == CONTENT_TYPE_OGG && !isOggTheora) || contentType == CONTENT_TYPE_WMA || contentType == CONTENT_TYPE_WAVPACK || contentType == CONTENT_TYPE_FLAC || contentType == CONTENT_TYPE_PCM || contentType == CONTENT_TYPE_AIFF || contentType == CONTENT_TYPE_APE || contentType == CONTENT_TYPE_MP4 || contentType == CONTENT_TYPE_DSD;
+    return contentType == CONTENT_TYPE_MP3 || (contentType == CONTENT_TYPE_OGG && !isOggTheora) || contentType == CONTENT_TYPE_WMA || contentType == CONTENT_TYPE_WAVPACK || contentType == CONTENT_TYPE_FLAC || contentType == CONTENT_TYPE_PCM || contentType == CONTENT_TYPE_AIFF || contentType == CONTENT_TYPE_APE || contentType == CONTENT_TYPE_MP4 || contentType == CONTENT_TYPE_DSD || contentType == CONTENT_TYPE_MKV;
 }
 
 void TagLibHandler::addField(
@@ -366,7 +373,7 @@ void TagLibHandler::makeComment(
 bool TagLibHandler::fillMetadata(const std::shared_ptr<CdsObject>& obj)
 {
     auto item = std::dynamic_pointer_cast<CdsItem>(obj);
-    if (!item || !isEnabled)
+    if (!item || !enabled)
         return false;
 
     auto res = item->getResource(ContentHandler::DEFAULT);
@@ -396,6 +403,8 @@ bool TagLibHandler::fillMetadata(const std::shared_ptr<CdsObject>& obj)
         extractAiff(fs, item, res);
     } else if (contentType == CONTENT_TYPE_PCM) {
         extractPcm(fs, item, res);
+    } else if (contentType == CONTENT_TYPE_MKV) {
+        extractMkv(fs, item, res);
     } else {
         log_warning("TagLibHandler: File '{}' has unhandled content type '{}'", item->getLocation().c_str(), contentType.c_str());
         result = false;
@@ -423,7 +432,7 @@ std::unique_ptr<IOHandler> TagLibHandler::serveContent(
     const std::shared_ptr<CdsResource>& resource)
 {
     auto item = std::dynamic_pointer_cast<CdsItem>(obj);
-    if (!item) // not streamable
+    if (!item && !enabled) // not streamable
         return nullptr;
 
     std::string contentType = getValueOrDefault(mimeContentTypeMappings, item->getMimeType());
@@ -567,6 +576,29 @@ std::unique_ptr<IOHandler> TagLibHandler::serveContent(
         const TagLib::ByteVector& data = pic->data();
 
         return std::make_unique<MemIOHandler>(data.data(), data.size());
+    }
+    if (contentType == CONTENT_TYPE_MKV) {
+#if TAGLIB_MAJOR_VERSION >= 2 && TAGLIB_MINOR_VERSION >= 2
+        auto mkv = TagLib::Matroska::File(&roStream);
+
+        if (!mkv.isValid()) {
+            throw_std_runtime_error("Could not open file {}: as Matroska file", itemLocation);
+        }
+        auto attachments = mkv.attachments();
+        if (attachments) {
+            for (auto attmt : attachments->attachedFileList()) {
+                auto fileName = attmt.fileName().to8Bit(true);
+                if (startswith(fileName, "cover")) {
+                    const TagLib::ByteVector& data = attmt.data();
+                    return std::make_unique<MemIOHandler>(data.data(), data.size());
+                }
+            }
+        }
+        throw_std_runtime_error("Matroska resource {} missing picture information", itemLocation);
+#else
+        log_warning("For Matroska support in file '{}' TaglibHandler needs to be built with taglib 2.2 and above", itemLocation);
+        return nullptr;
+#endif
     }
 
     throw_std_runtime_error("File {} has unsupported content type {}", itemLocation, contentType);
@@ -1003,6 +1035,38 @@ void TagLibHandler::extractPcm(
     setBitsPerSample(item, res, wav);
 #else
     log_warning("For PCM support in file '{}' TaglibHandler needs to be built with taglib 2 and above", item->getLocation().c_str());
+#endif
+}
+
+void TagLibHandler::extractMkv(
+    TagLib::IOStream& roStream,
+    const std::shared_ptr<CdsItem>& item,
+    const std::shared_ptr<CdsResource>& res) const
+{
+#if TAGLIB_MAJOR_VERSION >= 2 && TAGLIB_MINOR_VERSION >= 2
+    auto mkv = TagLib::Matroska::File(&roStream);
+
+    if (!mkv.isValid()) {
+        log_info("TagLibHandler {}: does not appear to be a valid Matroska file", item->getLocation().c_str());
+        return;
+    }
+    auto sc = converterManager->i2i();
+    populateGenericTags(item, res, mkv, mkv.properties(), sc);
+    auto attachments = mkv.attachments();
+    if (attachments) {
+        for (auto attmt : attachments->attachedFileList()) {
+            auto fileName = attmt.fileName().to8Bit(true);
+            if (startswith(fileName, "cover")) {
+                std::string artMimetype = attmt.mediaType().to8Bit(true);
+                addArtworkResource(item, ContentHandler::ID3, artMimetype);
+                log_debug("{} -> {}", fileName, artMimetype);
+            }
+        }
+    }
+
+    setBitsPerSample(item, res, mkv);
+#else
+    log_warning("For Matroska support in file '{}' TaglibHandler needs to be built with taglib 2.2 and above", item->getLocation().c_str());
 #endif
 }