+++ ReleaseNotes.md
@@ -6,11 +6,13 @@
 
 - Split config files into modules (from-file)
 - Export and import changed objects
+- Extract Matroska metadata also with taglib 2.2
+- Select content type for import library to reduce import time
 
 ### Code Improvements
 
 - Remove prefix for location type in database
-- Update maximum versions of npupnp (6.2.3), cmake (4.2.3), pupnp (1.14.30)
+- Update maximum versions of npupnp (6.2.3), cmake (4.2.3), pupnp (1.14.30), taglib (2.2)
 
 ## v3.1.1